refactor(page): extract editor modal organisms (phase 2)
Move the ~3.2k-line modal block out of page.tsx's JSX return into 5
domain-grouped organisms under components/organisms/modals/:
DiagnosticsModals, MeshRepairModals, ModifierModals, SceneFileModals,
PrintingModals. Each takes its Home state/handlers as identically-named
props (modal state stays in Home), so behavior is unchanged. Worktree
subagent, integrated + re-verified here.

page.tsx: 16,132 -> 13,198 lines (-2,934). Verified via `tsc --noEmit`.
Smoke-test the modals (diagnostics, mesh-repair, hollowing/hole-punch
reset, scene save/close, printing confirm/reslice) before merge.
